+++ src/Commands/EditCredentialsCommand.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace BeyondCode\Credentials;
+namespace RtoWebsites\Credentials\Commands;
 
 use Illuminate\Console\Command;
+use RtoWebsites\Credentials\Credentials;
+use RtoWebsites\Credentials\Exceptions\InvalidJSON;
 use Symfony\Component\Process\Process;
-use BeyondCode\Credentials\Exceptions\InvalidJSON;
 
 class EditCredentialsCommand extends Command
 {
@@ -25,8 +26,7 @@ class EditCredentialsCommand extends Command
     /**
      * The command handler.
      *
-     * @param \BeyondCode\Credentials\Credentials $credentials
-     * @return void
+     * @param Credentials $credentials
      */
     public function handle(Credentials $credentials)
     {
@@ -39,12 +39,7 @@ class EditCredentialsCommand extends Command
 
         fwrite($handle, json_encode($decrypted, JSON_PRETTY_PRINT | JSON_FORCE_OBJECT));
 
-        $editor = env('EDITOR', 'vi');
-
-        $process = new Process($editor.' '.$meta['uri']);
-
-        $process->setTty(true);
-        $process->mustRun();
+        $this->runEditor($meta['uri']);
 
         $data = json_decode(file_get_contents($meta['uri']), JSON_OBJECT_AS_ARRAY);
 
@@ -55,5 +50,17 @@ class EditCredentialsCommand extends Command
         $credentials->store($data, $filename);
 
         $this->info('Successfully updated credentials.');
+    }
+
+    /**
+     * @param string $argument
+     */
+    public function runEditor(string $argument): void
+    {
+        $editor = config('credentials.editor');
+
+        $process = new Process([$editor, $argument]);
+        $process->setTty(true);
+        $process->mustRun();
     }
 }