+++ tests/CredentialTest.php
@@ -1,161 +1,80 @@
 <?php
 
-namespace BeyondCode\Credentials\Tests;
+namespace RtoWebsites\Credentials\Tests;
 
+use Illuminate\Foundation\Testing\WithFaker;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Str;
-use Orchestra\Testbench\TestCase;
-use Illuminate\Encryption\Encrypter;
-use BeyondCode\Credentials\Credentials;
-use BeyondCode\Credentials\CredentialsServiceProvider;
+use RtoWebsites\Credentials\Credentials;
 
 class CredentialTest extends TestCase
 {
-    protected function getPackageProviders($app)
+    use WithFaker;
+
+    /**
+     * @test
+     */
+    public function credentials_are_stored_encrypted()
     {
-        return [CredentialsServiceProvider::class];
-    }
+        Config::set('credentials.key', $credentialsKey = Str::random(32));
 
-    public function tearDown(): void
-    {
-        @unlink(__DIR__ . '/temp/credentials.php.enc');
-    }
+        /* @var Credentials $credentials */
+        $credentials = $this->app->get(Credentials::class);
 
-    /** @test */
-    public function it_can_load_encrypted_files()
-    {
-        $masterKey = Str::random(16);
+        $this->assertFileNotExists($file = Config::get('credentials.file'));
+        $this->assertEmpty($credentials->load($file));
 
-        // create fake credentials
-        $encrypter = new Encrypter($masterKey);
-        $fakeCredentials = $encrypter->encrypt([
-            'key' => 'my-secret-value'
-        ]);
-
-        $encryptedData = '<?php return '.var_export($fakeCredentials, true).';';
-
-        file_put_contents(__DIR__ . '/temp/credentials.php.enc', $encryptedData);
-
-        $credentials = new Credentials($encrypter);
-
-        $credentials->load(__DIR__ . '/temp/credentials.php.enc');
-
-        $this->assertSame('my-secret-value', $credentials->get('key'));
-    }
-
-    /** @test */
-    public function it_can_store_data_encrypted()
-    {
-        $masterKey = Str::random(16);
-
-        $encrypter = new Encrypter($masterKey);
-
-        $data = [
-            'key' => 'my-secret-value'
+        $secretData = [
+            'key' => $secretPassword = $this->faker->password,
         ];
 
-        $credentials = new Credentials($encrypter);
+        $credentials->store($secretData, $file);
 
-        $credentials->store($data, __DIR__ . '/temp/credentials.php.enc');
+        $this->assertFileExists($file);
 
-        $this->assertSame([
-            'key' => 'my-secret-value'
-        ], $credentials->load(__DIR__ . '/temp/credentials.php.enc'));
-    }
+        $encryptedData = file_get_contents($file);
 
-    /** @test */
-    public function it_returns_decrypted_data()
-    {
-        $masterKey = Str::random(16);
+        $this->assertStringNotContainsString($secretPassword, $encryptedData);
+        $this->assertEquals($secretData, $credentials->load($file));
+        $this->assertEquals($secretPassword, $credentials->get('key'));
 
-        // create fake credentials
-        $encrypter = new Encrypter($masterKey);
-        $fakeCredentials = $encrypter->encrypt([
-            'key' => 'my-secret-value'
-        ]);
-
-        $encryptedData = '<?php return '.var_export($fakeCredentials, true).';';
-
-        file_put_contents(__DIR__ . '/temp/credentials.php.enc', $encryptedData);
-
-        $credentials = new Credentials($encrypter);
-
-        $decrypted = $credentials->load(__DIR__ . '/temp/credentials.php.enc');
-
-        $this->assertSame([
-            'key' => 'my-secret-value'
-        ], $decrypted);
     }
 
     /**
      * @test
-     * @expectedException Illuminate\Contracts\Encryption\DecryptException
      */
-    public function it_can_not_decrypt_with_the_wrong_key()
+    public function the_credentials_key_can_be_base_64_encoded()
     {
-        $masterKey = Str::random(16);
+        $credentialsKey = base64_encode(Str::random(32));
+        Config::set('credentials.key', "base64:$credentialsKey");
 
-        // create fake credentials
-        $encrypter = new Encrypter($masterKey);
-        $fakeCredentials = $encrypter->encrypt([
-            'key' => 'my-secret-value'
-        ]);
+        $secretData = [
+            'key' => $secretPassword = $this->faker->password,
+        ];
 
-        $encryptedData = '<?php return '.var_export($fakeCredentials, true).';';
+        /* @var Credentials $credentials */
+        $credentials = $this->app->get(Credentials::class);
+        $credentials->store($secretData, $file = Config::get('credentials.file'));
+        $loadedData = $credentials->load($file);
 
-        file_put_contents(__DIR__ . '/temp/credentials.php.enc', $encryptedData);
-
-        $credentials = new Credentials(new Encrypter(Str::random(16)));
-
-        $credentials->load(__DIR__ . '/temp/credentials.php.enc');
+        $this->assertEquals($secretData, $loadedData);
+        $this->assertEquals($secretPassword, $credentials->get('key'));
     }
 
-    /** @test */
+    /**
+     * @test
+     */
     public function it_can_use_the_helper_function()
     {
-        $this->app['config']->set('credentials.file', __DIR__ . '/temp/credentials.php.enc');
-
-        $data = [
-            'key' => 'my-secret-value'
+        $secretData = [
+            'key' => $secretPassword = $this->faker->password,
         ];
 
-        $credentials = app(Credentials::class);
+        /* @var Credentials $credentials */
+        $credentials = $this->app->get(Credentials::class);
+        $credentials->store($secretData, Config::get('credentials.file'));
 
-        $credentials->store($data, __DIR__ . '/temp/credentials.php.enc');
-
-        $this->assertSame('my-secret-value', credentials('key'));
+        $this->assertSame($secretPassword, credentials('key'));
     }
 
-    /** @test */
-    public function it_can_give_a_default_to_the_helper_function()
-    {
-        $this->app['config']->set('credentials.file', __DIR__ . '/temp/credentials.php.enc');
-
-        $data = [
-            'key' => 'my-secret-value'
-        ];
-
-        $credentials = app(Credentials::class);
-
-        $credentials->store($data, __DIR__ . '/temp/credentials.php.enc');
-
-        $this->assertSame('my-fallback-value', credentials('wrong-key', 'my-fallback-value'));
-    }
-
-    /** @test */
-    public function it_replaces_credential_strings_in_the_configuration_files()
-    {
-        $this->app['config']->set('credentials.file', __DIR__ . '/temp/credentials.php.enc');
-
-        $data = [
-            'key' => 'my-secret-value'
-        ];
-
-        $credentials = app(Credentials::class);
-
-        $credentials->store($data, __DIR__ . '/temp/credentials.php.enc');
-
-        $this->app['config']->set('credentials.secret', Credentials::CONFIG_PREFIX.'key');
-
-        $this->assertSame('my-secret-value', credentials('key'));
-    }
 }