+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to `laravel-credentials` will be documented in this file
 
+## 2.0.0
+
+### Added
+- configuration for the editor to be used with the `credentials:edit` command
+
+### Changed
+- change the general namespace to `RtoWebsites`
+- move commands and providers to distinct namespaces
+- increase overall test coverage
+
+### Fixed
+- for Laravel 6 the `credentials` helper can be used in config files again
+
 ## 1.0.0 - 2018-05-15
 
 - initial release